# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Catch API
  version: 0.1.0
  description: Backend contracts for Catch modular monolith.
servers:
  - url: /api/v1
...
paths:
...
  /articles/drafts:
    post:
      tags: [Articles]
      summary: Create article draft
      operationId: createArticleDraft
      security:
        - cookieSession: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDraftRequest"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/my:
    get:
      tags: [Articles]
      summary: List current user articles
      operationId: listMyArticles
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: after_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current user articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraftList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/drafts/{articleId}:
    get:
      tags: [Articles]
      summary: Get own article draft
      operationId: getArticleDraft
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Article draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Articles]
      summary: Update own article draft and create new revision
      operationId: updateArticleDraft
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDraftRequest"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/drafts/{articleId}/submit:
    post:
      tags: [Articles]
      summary: Submit draft for moderation or direct publication
      operationId: submitArticleDraft
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDraftRequest"
      responses:
        "200":
          description: Submitted draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/drafts/{articleId}/archive:
    post:
      tags: [Articles]
      summary: Archive own draft or moderation submission
      operationId: archiveArticleDraft
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Archived draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/{articleId}:
    get:
      tags: [Articles]
      summary: Get published article
      operationId: getPublishedArticle
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Published article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicArticle"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /feed:
    get:
      tags: [Articles]
      summary: List published feed articles
      operationId: listFeed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Feed articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
        "500":
          $ref: "#/components/responses/InternalError"
  /feed/popular:
    get:
      tags: [Articles]
      summary: List popular articles from the last 14 days
      operationId: listPopularFeed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Popular feed articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
        "500":
          $ref: "#/components/responses/InternalError"
  /articles/feed:
    get:
      tags: [Articles]
      summary: List personalized feed articles for current user
      operationId: listMyFeed
      security:
        - cookieSession: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Personalized feed articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /search:
    get:
      tags: [Articles]
      summary: Search published articles with PostgreSQL fallback search
      operationId: searchArticles
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 3
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: catch_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque cursor from previous list response.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user has no permission for this action
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ArticleDocument:
      type: object
      additionalProperties: true
      required: [type, version, blocks]
      properties:
        type:
          type: string
          const: catch.article
        version:
          type: integer
          minimum: 1
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
    CreateDraftRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        content:
          $ref: "#/components/schemas/ArticleDocument"
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 64
    UpdateDraftRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        content:
          $ref: "#/components/schemas/ArticleDocument"
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 64
    SubmitDraftRequest:
      type: object
      additionalProperties: false
      properties:
        publish_at:
          type: string
          format: date-time
    ArticleDraft:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - current_revision_id
        - moderation_required
        - title
        - content
        - excerpt
        - tags
        - version
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, in_moderation, ready_to_publish, published, archived, removed]
        current_revision_id:
          type: string
          format: uuid
        moderation_required:
          type: boolean
        title:
          type: string
        content:
          $ref: "#/components/schemas/ArticleDocument"
        excerpt:
          type: string
        cover_url:
          type: string
          format: uri-reference
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
          minimum: 1
        scheduled_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
    ArticleDraftList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArticleDraft"
    PublicArticle:
      type: object
      additionalProperties: false
      required: [id, author_id, title, content, excerpt, tags, reactions_up, reactions_down, reaction_score, published_at]
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          $ref: "#/components/schemas/ArticleDocument"
        excerpt:
          type: string
        cover_url:
          type: string
          format: uri-reference
        tags:
          type: array
          items:
            type: string
        reactions_up:
          type: integer
          minimum: 0
        reactions_down:
          type: integer
          minimum: 0
        reaction_score:
          type: integer
        published_at:
          type: string
          format: date-time
    ArticleList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArticleListItem"
        next_cursor:
          type: string
    ArticleListItem:
      type: object
      additionalProperties: false
      required: [id, author_id, title, excerpt, tags, reactions_up, reactions_down, reaction_score, published_at]
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        title:
          type: string
        excerpt:
          type: string
        cover_url:
          type: string
          format: uri-reference
        tags:
          type: array
          items:
            type: string
        reactions_up:
          type: integer
          minimum: 0
        reactions_down:
          type: integer
          minimum: 0
        reaction_score:
          type: integer
        published_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [type, title, status, code, message]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - not_implemented
            - service_unavailable
            - internal_error
        message:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true